pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news-kw"
version = "0.1.0"
description = "News keyword analysis pipeline for academic research"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "scikit-learn>=1.2.0",
    "nltk>=3.8",
    "pyyaml>=6.0",
    "matplotlib>=3.6.0",
    "networkx>=3.0",
    "tqdm>=4.65.0",
    "pyarrow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools]
packages = ["news_kw"]

[tool.setuptools.package-dir]
news_kw = "src/news_kw"

